requirements.txt
python-telegram-bot==7.0.1
Pyrebase==3.0.27
jwt==0.3.2
PyYAML==3.12
